Do not add title field to the navbar
For most services, no title is defined, so this is a no-op.

For those services where a title is defined (currently e.g. Zenodo and Rucio)
this adds the title to the navbar as a link to "/". This isn't useful — it
means that e.g. when deployed on sdc-dev these words  link to
https://sdc-dev.astron.nl/ which is a 404.